requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.24,<1.0
requests>=2.31,<3.0
httpx>=0.27,<1.0
python-multipart>=0.0.9,<1.0

# Database & ORM
tortoise-orm>=0.20,<1.0
asyncpg>=0.29,<1.0

# Migrations
aerich>=0.7,<1.0


